Dynamically allocate password instead of using VLA
This allows the code to compile with `cl` (MS Windows).

Additionally allows massive passwords to be made more easily I suppose,
as we aren't allocating (potentially dangerously) on the stack anymore.
But, generation may be slightly slower due to allocation of memory. We
could possibly have some statically allocated memory (e.g. 20 bytes) and
if we go over then malloc().